Traffic offenses are simplified by being separated into two charge types, by violation vs non-violation. The relevant behavior is:
 - All traffic of these are ineligible if convicted.
 - Traffic violations (for convictions and dismissals) are not considered in time analysis.
 - Dismissed traffic violations are ineligible, because they are never mentioned in the expungement law (nor dismissed civil offenses of any kind).
 - Dismissed traffic non-violations are eligible however, because they're criminal charges which are eligible.

All of this logic was already handled correctly. This change is only to better specify the Charge Type that is displayed to the user.

The unit tests are rearranged to match but are otherwise left unchanged, and they have inconsistencies in format. A larger overhaul of the unit test format for all charge types can handle this.
