+++ src/backend/expungeservice/models/charge_types/traffic_non_violation.py
@@ -1,0 +1,15 @@
+from dataclasses import dataclass
+
+from expungeservice.models.charge import Charge
+from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus
+
+
+@dataclass(eq=False)
+class TrafficNonViolation(Charge):
+    type_name: str = "Traffic Non-Violation"
+
+    def _default_type_eligibility(self):
+        if self.acquitted():
+            return TypeEligibility(EligibilityStatus.ELIGIBLE, reason="Dismissal eligible under 137.225(1)(b)")
+        else:
+            return TypeEligibility(EligibilityStatus.INELIGIBLE, reason="Ineligible under 137.225(7)(a)")
